Move `NumericPrefixWithSuffix` into utils
