+++ crates/project_panel/Cargo.toml
@@ -26,7 +26,7 @@ serde_json.workspace = true
 settings.workspace = true
 theme.workspace = true
 ui.workspace = true
-unicase = "2.6"
+unicase.workspace = true
 util.workspace = true
 client.workspace = true
 workspace.workspace = true
